requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
Pillow==10.0.0
opencv-python==4.8.0.74
albumentations==1.3.1
PyYAML==6.0
pytest==7.4.0
black==23.7.0
ruff==0.0.280
flake8==6.0.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
